Fix import of parse_requirements
Starting in pip 10.x, parse_requirements was moved to an internal
namespace to help demonstrate the fact that it is not a public API.

As a workaround until a broader decision can be made about how this
should look long term, this change attepts to perform the import using
the pip 10.x path and if it fails, falls back on the older path.

Fixes #80.