Fix check uwsgi options
Even if uwsgi is avaiable in sys.modules, you need to check if uwsgi.opt
is available. This code was broken in latest version of uwsgi when
running a process without uwsgi. The uwsgi documentation unfortunately is not clear about this https://uwsgi-docs.readthedocs.io/en/latest/PythonModule.html

Also the link for LaunchDarkly documentation is broken.