bugfix: Use notify_all
The standard library threading.Condition class had `notifyAll`
implemented as an alias to `notify_all` for backwards comaptibility with
older versions of python which still used the camel case methods.
Python3.10 deprecated the `notifyAll` alias.

This change removes the call to `notifyAll`, in favor of `notify_all`
which should exist as far back as python 3.5.