Exclude booleans when getting bucketable value (#181)
When calculating a bucket, we get the bucketable value from the
specified bucket by attribute. If this value is a string or an int, we
can use it. Otherwise, we return None.

Python considers a bool an instance of an int, which isn't what we want.
So we need to add an explicit exclusion for this.