Fix mypy type checking (#184)
A [customer requested][original-pr] that we start including a py.typed file in our
repository. This would enable mypy to take advantage of our typehints.
Unfortunately, this didn't completely solve the customers issue.

A [second pr][second-pr] was opened to address the missing step of including the
py.typed file in the `Manifest.in` file. However, this change alone is not
sufficient.

According to the [documentation][include_package_data], you must also
include the `include_package_data=True` directive so that files
specified in the `Manifest.in` file are included in distribution.

[original-pr]: https://github.com/launchdarkly/python-server-sdk/pull/166
[second-pr]: https://github.com/launchdarkly/python-server-sdk/pull/172
[include_package_data]: https://setuptools.pypa.io/en/latest/userguide/datafiles.html#include-package-data