Upgrade pip to fix failing CI build (#216)
The CI build was failing because pip had an outdated list of available wheels for installation. Since it couldn't find a match, it was trying to build a package from source, which requires the rust compiler, which in turn isn't present on some of the docker images.

By updating pip we get the updated list of available wheels, thereby allowing us to bypass source building and the need for the rust compiler entirely.