Fix unsafe access to flag's trackEvents field (#218)
When using the test data source, the trackEvents may not be required.
Additionally, when receiving this information from LD, there is no
guarantee that field will be provided. Instead of assuming that field
always exists on the flag model, we should instead retrieve it with a
default value to prevent index errors.