docs: Fix sphinx documentation failures (#219)
Our documentation file had a configuration issue, which emits a warning
when building with sphinx. This caused the build process on
readthedocs.org to fail because they treat warnings as errors.

We didn't catch this because:

1. Warnings are treated like warnings when generating documentation
   locally.
2. We don't try to generate the documentation as part of our CI process.

This commit should resolve both the underlying issue and the
deficiencies that prevented us from catching this initially.