fix: Move tests under ldclient namespace (#29) (#283)
Previous distributions of this package included two packages -- ldclient
and testing. This top level testing namespace can conflict with other
packages. In fact, it conflicts with our own eventsource library.

In general this doesn't matter, but it may if:

1. You are using a build process that warns about conflicts (see [this
issue][1])
2. You want to install the sdist on an unsupported platform and would
like to be able to verify the tests.

To resolve this issue, we are moving the testing folder into the
ldclient package. These testing files will only be included in the sdist
format. This allows for a smaller wheel size while also allowing for
flexibility with consumers.

[1]: launchdarkly/python-server-sdk#281