style: Fix and enforce formatting (#316)
This commit introduces both pycodestyle and isort tooling to the CI
check pipeline. These tools ensure a consistent code style and import
order across the codebase.

The commit also includes a number of fixes to the codebase to ensure it
passes the new checks. Most of these were automated through the use of
the CLI tool `black`. The rest were fixed manually.